requirements.txt
# Flask Web Framework
Flask==3.1.0
Flask-Cors==5.0.0
Werkzeug==3.1.3

# Deep Learning Libraries
tensorflow==2.17.1

# Image Processing
opencv-python==4.10.0.84
opencv-contrib-python==4.10.0.84
Pillow==11.0.0
scikit-image==0.24.0
imageio==2.36.0

# Scientific Computing
numpy==1.26.4
scipy==1.13.1
matplotlib==3.9.2

# Data Processing
h5py==3.12.1
PyYAML==6.0.2

# Utilities
requests==2.32.3
python-dateutil==2.9.0.post0
pytz==2024.2
tqdm==4.66.5

# 3D Modeling
trimesh==4.5.3
rtree==1.3.0

# Google Generative AI (Gemini) - v0.8+ with new SDK
google-genai==0.8.0
google-ai-generativelanguage==0.6.1

# Environment variables
python-dotenv==1.0.1
